provider/pyproject.toml
[project]
name = "provider"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.31.0",
    "email-validator>=2.3.0",
    "fastapi[standard]>=0.136.0",
    "pgvector>=0.4.2",
    "pydantic>=2.13.2",
    "sqlalchemy[asyncio]>=2.0.49",
    "structlog>=25.5.0",
]
